+++ noodles-vcf/examples/vcf_reheader.rs
@@ -25,7 +25,7 @@ fn add_comment(header: &mut vcf::Header) -> Result<(), Box<dyn std::error::Error
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
-    let mut reader = vcf::reader::Builder::default().build_from_path(src)?;
+    let mut reader = vcf::io::reader::Builder::default().build_from_path(src)?;
 
     let mut header = reader.read_header()?;
     add_comment(&mut header)?;