+++ noodles-vcf/examples/vcf_view.rs
@@ -12,7 +12,7 @@ use noodles_vcf as vcf;
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
-    let mut reader = vcf::reader::Builder::default().build_from_path(src)?;
+    let mut reader = vcf::io::reader::Builder::default().build_from_path(src)?;
     let header = reader.read_header()?;
 
     let stdout = io::stdout().lock();