+++ noodles-vcf/src/lib.rs
@@ -10,7 +10,7 @@
 //! ```no_run
 //! use noodles_vcf as vcf;
 //!
-//! let mut reader = vcf::reader::Builder::default().build_from_path("sample.vcf")?;
+//! let mut reader = vcf::io::reader::Builder::default().build_from_path("sample.vcf")?;
 //! let header = reader.read_header()?;
 //!
 //! for result in reader.records(&header) {
@@ -24,19 +24,17 @@
 mod r#async;
 
 pub mod header;
-pub mod indexed_reader;
 mod indexer;
 pub mod io;
 pub mod lazy;
-pub mod reader;
 pub mod record;
 mod variant_reader;
 mod variant_writer;
 pub mod writer;
 
 pub use self::{
-    header::Header, indexed_reader::IndexedReader, indexer::index, reader::Reader, record::Record,
-    variant_reader::VariantReader, variant_writer::VariantWriter, writer::Writer,
+    header::Header, indexer::index, record::Record, variant_reader::VariantReader,
+    variant_writer::VariantWriter, writer::Writer,
 };
 
 #[cfg(feature = "async")]